+++ db/migrate/20160919220802_create_payments.rb
@@ -1,7 +1,7 @@
 class CreatePayments < ActiveRecord::Migration
   def change
     create_table :payments do |t|
-      t.integer :cost
+      t.decimal :cost
       t.string :store
       t.integer :purchase_id
